+++ app/src/main/java/me/zchang/onchart/student/LabelCourse.java
@@ -42,7 +42,7 @@ public class LabelCourse extends Course implements Parcelable{
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(this.id);
+        dest.writeLong(this.id);
         dest.writeString(this.name);
         dest.writeString(this.teacher);
         dest.writeString(this.classroom);
@@ -57,7 +57,7 @@ public class LabelCourse extends Course implements Parcelable{
         @Override
         public Course createFromParcel(Parcel source) {
             Course course = new LabelCourse();
-            course.setId(source.readInt());
+            course.setId(source.readLong());
             course.setName(source.readString());
             course.setTeacher(source.readString());
             course.setClassroom(source.readString());
